Add Flags middleware (#9)
This adds a new `middleware` package as well as `middleware.Flags`,
which allows you to supply a callback function receiving a
`flag.FlagSet` to define flags on.

In order to make this work, middleware is now applied before route
matching. Otherwise, middleware can't change the route matching, which
is necessary, because flags are part of the route before parsing.

I had to disable `Router.Scope` because I can't currently make it work
with the middleware changes, and I'm prioritizing the flags feature. See
#8.

Fixes #4